Move minify utils to separate folder
